requirements.txt
scapy>=2.7.0
pyyaml>=6.0
matplotlib>=3.5
psutil>=5.9
os-ken>=2.6
